requirement.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
docutils==0.20.1
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
lockfile==0.12.2
msgspec==0.18.5
protobuf==4.25.1
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
python-daemon==3.0.1
python-dateutil==2.8.2
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
smmap==5.0.1
typing_extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
zeroapi==0.5.0
